feat(http-client): Set User-Agent header dynamically
This commit updates the HTTP client code to set the 'User-Agent' header dynamically based on the SDK version. The User-Agent now follows the format 'FlutterwavePHP/{SDK_VERSION}', where {SDK_VERSION} is the version of the SDK obtained from the EnvVariables class.This change improves the clarity of user-agent information for HTTP requests, helping with tracking and debugging.