Update Configure.php
When a custom `env` variable is created, the bootstrap method was still trying to find the default env keys. 

With this new addition, if a user has used the `PackageConfig::setUp` method already to set their custom `env` variables, bootstrap would not have assume `self::$config` is not set again.